refactor(layout): move away from mdsvex system (project)
