+++ src/content/project/blossom/index.ru.svx
@@ -5,7 +5,6 @@
 	"draft": true,
 	"featured": true,
 	"lang": "ru",
-	"layout": "project",
 	"maintained": false,
 	"npm": "@ericrovell/blossom",
 	"repository": "EricRovell/blossom",
