+++ src/content/project/project-euler/index.ru.svx
@@ -5,7 +5,6 @@
 	"draft": true,
 	"featured": false,
 	"lang": "ru",
-	"layout": "project",
 	"maintained": true,
 	"repository": "EricRovell/project-euler-ts",
 	"title": "project-euler",
