+++ src/content/project/radix/index.ru.svx
@@ -4,7 +4,6 @@
 	"description": "Radix — библиотека для перевода чисел в различные системы счисленя, написанная на TypeScript.",
 	"featured": true,
 	"lang": "ru",
-	"layout": "project",
 	"maintained": true,
 	"npm": "@ericrovell/radix",
 	"repository": "EricRovell/radix",
