+++ src/content/project/rational/index.ru.svx
@@ -5,7 +5,6 @@
 	"draft": true,
 	"featured": true,
 	"lang": "ru",
-	"layout": "project",
 	"maintained": true,
 	"npm": "@ericrovell/rational",
 	"repository": "EricRovell/rational",
