+++ src/content/project/svelte-media-observer/index.en.svx
@@ -5,7 +5,6 @@
 	"draft": true,
 	"featured": true,
 	"lang": "en",
-	"layout": "project",
 	"maintained": true,
 	"npm": "svelte-media-observer",
 	"repository": "EricRovell/svelte-media-observer",
