+++ src/content/project/svelte-media-observer/index.ru.svx
@@ -5,7 +5,6 @@
 	"draft": true,
 	"featured": true,
 	"lang": "ru",
-	"layout": "project",
 	"maintained": true,
 	"npm": "svelte-media-observer",
 	"title": "svelte-media-observer",
