+++ src/lib/layout/index.ts
@@ -3,3 +3,4 @@ export { LayoutPage } from "./page-header";
 export { default as BlogpostLayout } from "./blogpost/blogpost.layout.svelte";
 export { default as GalleryItemLayout } from "./gallery/gallery.layout.svelte";
 export { default as PageLayout } from "./page/page.layout.svelte";
+export { default as ProjectLayout } from "./project/project.layout.svelte";
