+++ src/lib/layout/project/project.layout.module.css
@@ -1,3 +1,72 @@
+.header {
+	position: relative;
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	justify-content: center;
+	width: min(100% - var(--space-3), 1440px);
+	text-align: center;
+
+	& > h1 {
+		font-family: var(--font-secondary);
+		text-align: center;
+		margin-block: 0 var(--space-1-rel);
+		max-inline-size: 65ch;
+	}
+
+	& > p {
+		font-size: var(--font-size-1);
+		text-align: center;
+		max-inline-size: 45ch;
+		line-height: 1.4;
+		margin-block-end: var(--space-1-rel);
+
+		&:last-of-type {
+			margin-block-end: var(--space-4-rel);
+			font-size: 1em;
+			color: var(--surface-1-text-less-contrast);
+		}
+	}
+
+	& > footer {
+		display: contents;
+		font-size: var(--font-size-1);
+	}
+}
+
+.keywords {
+	display: flex;
+	flex-wrap: wrap;
+	justify-content: center;
+	gap: var(--space-2-rel);
+	color: var(--color-info);
+	margin-block-end: var(--space-3-rel);
+
+	& li {
+		padding: var(--space-2-rel) var(--space-3-rel);
+		border-radius: var(--radius-s);
+		font-size: var(--font-size-0);
+	}
+}
+
+.links {
+	--icon-size: 1.5em;
+
+	display: grid;
+	grid-auto-flow: column;
+	grid-auto-columns: 1fr;
+	justify-content: space-evenly;
+	gap: var(--space-4-rel);
+
+	& a {
+		display: flex;
+		flex-direction: column;
+		place-items: center;
+		gap: var(--space-2-rel);
+		font-size: var(--font-size-0);
+	}
+}
+
 .article {
 	grid-area: article;
 	display: grid;
